+++ client/src/pages/MessagePage/MessageDetail/styles.ts
@@ -3,20 +3,6 @@ import { css } from '@emotion/react';
 import { COLORS } from 'styles/colors';
 import { COMMON_SIZE } from 'styles/sizes';
 
-export const goBackButtonStyle = css({
-  padding: 0,
-  width: 20,
-  height: 20,
-  marginRight: 10,
-
-  ' svg': {
-    width: 20,
-    height: 20,
-    transform: 'rotateZ(90deg)',
-    fill: COLORS.PRIMARY_2,
-  },
-});
-
 export const messageDetailListStyle = css({
   width: '100%',
   flex: 1,
@@ -41,4 +27,18 @@ export const messageInputStyle = css({
   padding: '10px 10px',
   border: `1px solid ${COLORS.BOX_BORDER}`,
   borderRadius: COMMON_SIZE.BORDER_RADIUS,
+});
+
+export const goBackButtonStyle = css({
+  padding: 0,
+  width: 20,
+  height: 20,
+  marginRight: 10,
+
+  ' svg': {
+    width: 20,
+    height: 20,
+    transform: 'rotateZ(90deg)',
+    fill: COLORS.PRIMARY_2,
+  },
 });
