+++ client/src/types/message.d.ts
@@ -1,11 +1,12 @@
 export interface MessageMetaDataType {
   with: string;
+  username: string;
   lastCheckTime: string;
 }
 
-export interface MessageListType {
-  messages: MessageMetaDataType[];
-  lastPageConnectTime: string;
+export interface MessageDetailType {
+  contents: SingleMessageType[];
+  toUsername: string;
 }
 
 export interface SingleMessageType {
